Adds a new file that includes mesh utility functions
